requirements.txt
Django>=4.2.0
python-telegram-bot==20.7
aiogram>=3.0.0
python-dotenv>=1.0.0
Pillow==10.2.0
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
celery>=5.3.0
redis>=4.5.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0 